add jwt for category class

+++ task-backend/src/main/java/lk/ijse/gdse/taskbackend/controllers/CategoryController.java
@@ -1,8 +1,10 @@
 package lk.ijse.gdse.taskbackend.controllers;
 
+import lk.ijse.gdse.taskbackend.auth.JwtTokenProvider;
 import lk.ijse.gdse.taskbackend.dto.CategoryDTO;
 import lk.ijse.gdse.taskbackend.entity.Category;
 import lk.ijse.gdse.taskbackend.service.CategoryService;
+import lk.ijse.gdse.taskbackend.service.InventoryService;
 import lk.ijse.gdse.taskbackend.util.ResponseUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -14,15 +16,20 @@ import java.util.Optional;
 @CrossOrigin("*")
 @RequestMapping("/api/categories")
 public class CategoryController {
-    @Autowired
-    private final CategoryService categoryService;
 
-    public CategoryController(CategoryService categoryService) {
+    private final CategoryService categoryService;
+    private final InventoryService inventoryService;
+    private final JwtTokenProvider jwtTokenProvider;
+
+    @Autowired
+    public CategoryController(CategoryService categoryService, InventoryService inventoryService, JwtTokenProvider jwtTokenProvider) {
         this.categoryService = categoryService;
+        this.inventoryService = inventoryService;
+        this.jwtTokenProvider = jwtTokenProvider;
     }
 
     @PostMapping
-    public ResponseUtil saveCategory(@RequestBody CategoryDTO categoryDTO) {
+    public ResponseUtil saveCategory(@RequestBody CategoryDTO categoryDTO, @RequestHeader("Authorization") String authorizationHeader) {
         System.out.println(categoryDTO);
         Category category = categoryService.saveCategory(categoryDTO);
         return new ResponseUtil(200, "successfully save category", category);
@@ -30,42 +37,51 @@ public class CategoryController {
     }
 
     @GetMapping
-    public List<Category> getAllCategoies() {
-        List<Category> allCategories = categoryService.getAllCategories();
+    public List<Category> getAllCategoies(@RequestHeader("Authorization") String authorizationHeader) {
+        if (jwtTokenProvider.validateToken(authorizationHeader)) {
+            return categoryService.getAllCategories();
+        }
 //        return new ResponseUtil(200, "found all category", allCategories);
-        return allCategories;
+        return null;
     }
 
     @DeleteMapping("/{id}")
-    public ResponseUtil deleteCategory(@PathVariable Long id) {
-        categoryService.deleteCategory(id);
-        return new ResponseUtil(200, "Successfully deleted category", null);
+    public ResponseUtil deleteCategory(@PathVariable Long id, @RequestHeader("Authorization") String authorizationHeader) {
+        if (jwtTokenProvider.validateToken(authorizationHeader)) {
+            categoryService.deleteCategory(id);
+            return new ResponseUtil(200, "Successfully deleted category", null);
+        }
+        return new ResponseUtil(401, "UnAuthorization", null);
     }
 
-    @GetMapping(  "/findById/{id}")
-    public Category findCategoryById(@PathVariable Long id) {
-        Optional<Category> categoryById = categoryService.findCategoryById(id);
-        //return new ResponseUtil(200, "successfully find category", categoryById);
-        return categoryById.get();
+    @GetMapping("/findById/{id}")
+    public Category findCategoryById(@PathVariable Long id, @RequestHeader("Authorization") String authorizationHeader) {
+        if (jwtTokenProvider.validateToken(authorizationHeader)) {
+            Optional<Category> categoryById = categoryService.findCategoryById(id);
+            //return new ResponseUtil(200, "successfully find category", categoryById);
+            return categoryById.get();
+        }
+        return null;
     }
 
     @PutMapping
-    public ResponseUtil updateCategory(@RequestBody CategoryDTO categoryDTO) {
-        System.out.println(categoryDTO);
-        System.out.println("ok 1");
-        Category category = categoryService.updateCategory(categoryDTO);
-        System.out.println("ok 2");
-        if (category != null) {
-            System.out.println("ok 3");
-            return new ResponseUtil(200, "successfully update category", category);
-        } else {
-            return new ResponseUtil(200, "not found category", null);
+    public ResponseUtil updateCategory(@RequestBody CategoryDTO categoryDTO, @RequestHeader("Authorization") String authorizationHeader) {
+        if (jwtTokenProvider.validateToken(authorizationHeader)) {
+            Category category = categoryService.updateCategory(categoryDTO);
+            if (category != null) {
+                return new ResponseUtil(200, "successfully update category", category);
+            } else {
+                return new ResponseUtil(200, "not found category", null);
+            }
         }
-
+        return new ResponseUtil(401, "UnAuthorization", null);
     }
 
     @GetMapping(path = "/getIds")
-    public List<String> getAllNames() {
-        return categoryService.getAllCategoryIds();
+    public List<String> getAllNames(@RequestHeader("Authorization") String authorizationHeader) {
+        if (jwtTokenProvider.validateToken(authorizationHeader)) {
+            return categoryService.getAllCategoryIds();
+        }
+        return null;
     }
 }
